refactor(ci): finish ci_integration shim reduction (phase-3 closure)
Extract the two remaining in-file slices from phi_scan/ci_integration.py
into dedicated modules and reduce the shim to a pure compatibility
façade.

- Move comment-body formatting (BaselineComparison, build_comment_body,
  build_comment_body_with_baseline, _insert_baseline_context_into_comment,
  and comment-formatting constants) to phi_scan/ci/comment_body.py.
- Move orchestration dispatch (post_pr_comment, post_pull_request_comment,
  set_commit_status) to phi_scan/ci/dispatch.py.
- Reduce ci_integration.py from 361 to 117 lines: re-exports, explicit
  __all__, PRContext/get_pr_context aliases, updated docstring.
- Remove the stale docstring claim that platform-specific extras "remain
  here pending a follow-up migration PR"; those were already extracted.
- Remove four dead section banners at the bottom of the old shim.

Parity tests (tests/test_ci_integration_shim_parity.py):
- Symbol identity for every moved symbol (both slices).
- One behavioral smoke test for build_comment_body.
- One behavioral smoke test for post_pr_comment with monkeypatched
  adapter resolution.

No behavior changes. No unrelated refactors. Verified
BaseCIAdapter.import_findings_to_security_hub is the abstract interface
contract only (raises UnsupportedOperation); real implementation lives
in phi_scan/ci/aws_security_hub.py.
